+++ onnx_tf/backend_v4.py
@@ -1,0 +1,22 @@
+"""Backend for running ONNX on Tensorflow
+
+To run this, you will need to have Tensorflow installed as well.
+"""
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+from onnx_tf.backend import TensorflowBackendBase
+import tensorflow as tf
+
+class TensorflowBackend(TensorflowBackendBase):
+  """ Tensorflow Backend for ONNX
+  """
+
+  @classmethod
+  def handle_concat(cls, node, input_dict):
+    values = [input_dict[a] for a in node.inputs]
+    assert "axis" in node.attrs, ("axis is required.")
+    axis = node.attrs["axis"]
+    return [tf.concat(values, axis=axis)]
