+++ test/backend/test_node.py
@@ -610,10 +610,11 @@ class TestNode(unittest.TestCase):
                                    np.power(x, y))
 
   def test_reshape(self):
-    node_def = helper.make_node("Reshape", ["X"], ["Y"], shape=[10, 10])
+    node_def = helper.make_node("Reshape", ["X", "Y"], ["Z"])
     x = self._get_rnd(100)
-    output = run_node(node_def, [x])
-    np.testing.assert_almost_equal(output["Y"], x.reshape([10, 10]))
+    y = [10, 10]
+    output = run_node(node_def, [x, y])
+    np.testing.assert_almost_equal(output["Z"], x.reshape([10, 10]))
 
   def test_selu(self):
     node_def = helper.make_node("Selu", ["X"], ["Y"])
